Fixes for Klocwork issues

+++ src/benchmarks/benchmark.hpp
@@ -81,19 +81,19 @@ struct benchmark;
  * clos, nclos and opt_size in benchmark_info structure.
  */
 struct benchmark_args {
-	const char *fname;	 /* path to test file */
-	size_t fsize;		   /* size of test file */
-	bool is_poolset;	   /* test file is a poolset */
-	mode_t fmode;		   /* test file's permissions */
-	unsigned n_threads;	/* number of working threads */
-	uint64_t n_ops_per_thread; /* number of operations per thread */
-	bool thread_affinity;      /* set worker threads CPU affinity mask */
-	size_t dsize;		   /* data size */
-	unsigned seed;		   /* PRNG seed */
-	unsigned repeats;	  /* number of repeats of one scenario */
-	unsigned min_exe_time;     /* minimal execution time */
-	bool help;		   /* print help for benchmark */
-	void *opts;		   /* benchmark specific arguments */
+	const char *fname;       /* path to test file */
+	size_t fsize;		 /* size of test file */
+	bool is_poolset;	 /* test file is a poolset */
+	mode_t fmode;		 /* test file's permissions */
+	unsigned n_threads;      /* number of working threads */
+	size_t n_ops_per_thread; /* number of operations per thread */
+	bool thread_affinity;    /* set worker threads CPU affinity mask */
+	size_t dsize;		 /* data size */
+	unsigned seed;		 /* PRNG seed */
+	unsigned repeats;	/* number of repeats of one scenario */
+	unsigned min_exe_time;   /* minimal execution time */
+	bool help;		 /* print help for benchmark */
+	void *opts;		 /* benchmark specific arguments */
 };
 
 /*
